TPM: handle device reconfigurations
This patch looks at device changes due to modified XML. 
With the TPM device we have to handle cases where the TPM is removed
from the VM or its type or encryption mode are changed. Particularly
for the encryption we want to prevent the TPM encryption mode to be
changed and the TPM not working anymore since the passphrase it not
supplied.