Add support for TPM state encryption
Add support for having the TPM's state encrypted. This patch
builds on existing infrastructure for secrets and reuses the
XML encryption node used for QCoW2 images for example.

The follow XML for TPM devices is now support:

  <tpm model='tpm-tis'>
    <backend type='cuse-tpm'>
      <encryption format='default'/>
    </backend>
  </tpm>

FIXME: Users should not be able to add encryption to a cuse-tpm
if the TPM has already been started before; the code unfortunately
does not know whether parsing is done due to an update or startup
of libvirt.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>