tpm: Handle TPM emulator device reconfigurations
This patch looks at device changes due to modified XML. With the TPM
emulator we have to handle cases where the TPM is removed from the VM
or when it is reconfigured in some way.

In the device removal case we remove the storage path.

The function we create here will be used later to handle other device
changes as well.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>