tpm: Add support for choosing emulation of a TPM 2
This patch extends the TPM's device XML with TPM 2 support. This only works
for the emulator type backend and looks as follows:

    <tpm model='tpm-tis'>
      <backend type='emulator' tpmversion='2'/>
    </tpm>

Once the version of a TPM has been chosen it cannot be changed anymore unless
one removes the TPM device first and then reads it. However, one looses all
the secrets stored inside or tied to the emulated TPM by doing this.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>